Moves RequestInterceptor class into its own module
